requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pytest
httpx
python-multipart
pytest-asyncio
psycopg2-binary